Alpha 1.0
Version Alpha 1.0: In this version, a new work block for villagers: nether table and a new profession, the nether resident, have been added. This resident cannot appear in villages, but it can appear in a separate structure locked in a cage. The name of this structure is Nether House. Additionally, tin ore, tin ingot, and tin block have been added. By combining tin and copper, you can obtain bronze. Bronze can be used to create armor and tools that are slightly better than iron tools.